Refactor + Fix docs
