+++ Cargo.toml
@@ -10,9 +10,13 @@ documentation = "https://docs.rs/axum-password-worker/"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[features]
+default = ["bcrypt"]
+
 [dependencies]
-bcrypt = "0.14"
 tokio = { version = "1", features = ["full"] }
 rayon = "1"
 crossbeam-channel = "0.5"
 thiserror = "1"
+
+bcrypt = { version = "0.14", optional = true }
