+++ src/hasher_impls/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(feature = "bcrypt")]
+pub mod bcrypt;
