+++ src/lib.rs
@@ -31,184 +31,18 @@
 //! # Ok(())
 //! # }
 //! ```
-use rayon::ThreadPoolBuilder;
-use tokio::sync::oneshot;
+//!
+//! # Available feature flags
+//!
+//! There are some implementations included in the library. Each is tied to optional dependency features.
+//! * `bcrypt` - (default) (dependency), exports the Bcrypt And BcryptConfig types.
 
-/// Errors that can occur in the `PasswordWorker`.
-#[derive(Debug, thiserror::Error)]
-pub enum PasswordWorkerError<H: Hasher> {
-    #[error("Hashing error: {0}")]
-    Hashing(H::Error),
-    #[error("Channel send error: {0}")]
-    ChannelSend(#[from] crossbeam_channel::SendError<WorkerCommand<H>>),
-    #[error("Channel receive error: {0}")]
-    ChannelRecv(#[from] tokio::sync::oneshot::error::RecvError),
-    #[error("ThreadPool build error: {0}")]
-    ThreadPool(#[from] rayon::ThreadPoolBuildError),
-    #[error("No tokio runtime error: {0}")]
-    Runtime(#[from] tokio::runtime::TryCurrentError),
-}
+mod hasher;
+mod hasher_impls;
+mod worker;
 
-#[derive(Debug)]
-pub enum WorkerCommand<H: Hasher> {
-    Hash(
-        String,
-        H::Config,
-        oneshot::Sender<Result<String, PasswordWorkerError<H>>>,
-    ),
-    Verify(
-        String,
-        String,
-        oneshot::Sender<Result<bool, PasswordWorkerError<H>>>,
-    ),
-}
-
-pub trait Hasher: 'static {
-    type Config: Send + Sync + 'static;
-    type Error: std::error::Error + Send + Sync + 'static;
-    fn hash(data: impl AsRef<[u8]>, config: &Self::Config) -> Result<String, Self::Error>;
-    fn verify(data: impl AsRef<[u8]>, hash: &str) -> Result<bool, Self::Error>;
-}
-
-#[derive(Clone, Copy, Debug)]
-pub enum Bcrypt {}
-
-impl Hasher for Bcrypt {
-    type Config = BcryptConfig;
-    type Error = bcrypt::BcryptError;
-
-    fn hash(data: impl AsRef<[u8]>, config: &Self::Config) -> Result<String, Self::Error> {
-        bcrypt::hash(data, config.cost)
-    }
-
-    fn verify(data: impl AsRef<[u8]>, hash: &str) -> Result<bool, Self::Error> {
-        bcrypt::verify(data, hash)
-    }
-}
-
-#[derive(Clone, Copy)]
-pub struct BcryptConfig {
-    pub cost: u32,
-}
-
-/// A worker that handles password hashing and verification using a `rayon` thread pool
-/// and `crossbeam-channel`.
-///
-/// The `PasswordWorker` struct provides asynchronous password hashing and verification
-/// operations.
-#[derive(Debug, Clone)]
-pub struct PasswordWorker<H: Hasher> {
-    sender: crossbeam_channel::Sender<WorkerCommand<H>>,
-}
-
-impl<H: Hasher> PasswordWorker<H> {
-    /// Creates a new `PasswordWorker` with the given bcrypt cost and maximum number of threads.
-    ///
-    /// The `cost` parameter determines the computational cost of the bcrypt hashing algorithm.
-    /// The `max_threads` parameter specifies the maximum number of threads the worker can use.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// # #[tokio::main]
-    /// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// use axum_password_worker::{Bcrypt, PasswordWorker};
-    ///
-    /// let max_threads = 4; // rayon thread pool max threads
-    /// let password_worker: PasswordWorker<Bcrypt> = PasswordWorker::new(max_threads)?;
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub fn new(max_threads: usize) -> Result<Self, PasswordWorkerError<H>> {
-        let (sender, receiver) = crossbeam_channel::unbounded::<WorkerCommand<H>>();
-
-        let thread_pool = ThreadPoolBuilder::new().num_threads(max_threads).build()?;
-
-        tokio::runtime::Handle::try_current()?.spawn_blocking(move || {
-            while let Ok(command) = receiver.recv() {
-                match command {
-                    WorkerCommand::Hash(password, cost, result_sender) => {
-                        let result = thread_pool.install(|| H::hash(&password, &cost));
-                        result_sender
-                            .send(result.map_err(PasswordWorkerError::Hashing))
-                            .ok()?;
-                    }
-                    WorkerCommand::Verify(password, hash, result_sender) => {
-                        let result = thread_pool.install(|| H::verify(&password, &hash));
-                        result_sender
-                            .send(result.map_err(PasswordWorkerError::Hashing))
-                            .ok()?;
-                    }
-                }
-            }
-            Some(())
-        });
-
-        Ok(PasswordWorker { sender })
-    }
-
-    /// Asynchronously hashes the given password using bcrypt.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # #[tokio::main]
-    /// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// use axum_password_worker::{Bcrypt, BcryptConfig, PasswordWorker};
-    ///
-    /// let password = "hunter2";
-    /// let cost = 12; // bcrypt cost value
-    /// let max_threads = 4; // rayon thread pool max threads
-    /// let password_worker = PasswordWorker::<Bcrypt>::new(max_threads)?;
-    ///
-    /// let hashed_password = password_worker.hash(password, BcryptConfig { cost }).await?;
-    /// println!("Hashed password: {:?}", hashed_password);
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub async fn hash(
-        &self,
-        password: impl Into<String>,
-        cost: H::Config,
-    ) -> Result<String, PasswordWorkerError<H>> {
-        let (tx, rx) = oneshot::channel();
-
-        self.sender
-            .send(WorkerCommand::Hash(password.into(), cost, tx))?;
-
-        rx.await?
-    }
-
-    /// Asynchronously verifies a password against a bcrypt hash.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # #[tokio::main]
-    /// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// use axum_password_worker::{Bcrypt, BcryptConfig, PasswordWorker};
-    ///
-    /// let password = "hunter2";
-    /// let cost = 12; // bcrypt cost value
-    /// let max_threads = 4; // rayon thread pool max threads
-    /// let password_worker = PasswordWorker::<Bcrypt>::new(max_threads)?;
-    /// let hashed_password = password_worker.hash(password, BcryptConfig { cost }).await?;
-    ///
-    /// let is_valid = password_worker.verify(password, hashed_password).await?;
-    /// println!("Verification result: {:?}", is_valid);
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub async fn verify(
-        &self,
-        password: impl Into<String>,
-        hash: impl Into<String>,
-    ) -> Result<bool, PasswordWorkerError<H>> {
-        let (tx, rx) = oneshot::channel();
-
-        self.sender
-            .send(WorkerCommand::Verify(password.into(), hash.into(), tx))?;
-
-        rx.await?
-    }
-}
+pub use hasher::Hasher;
+#[cfg(feature = "bcrypt")]
+#[cfg_attr(docsrs, doc(cfg(feature = "bcrypt")))]
+pub use hasher_impls::bcrypt::{Bcrypt, BcryptConfig};
+pub use worker::{PasswordWorker, PasswordWorkerError};
